refactor(user) : Split user service from path
