--- src/main/java/atdd/path/domain/User.java
@@ -1,7 +1,0 @@
-package atdd.path.domain;
-
-public class User {
-  private Long id;
-  private String email;
-  private String name;
-}
