+++ src/main/java/atdd/user/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package atdd.path.repository;
+package atdd.user.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import atdd.path.entity.User;
+import atdd.user.entity.User;
 
 public interface UserRepository extends CrudRepository<User, Long> {
 }
