+++ src/test/java/atdd/user/TestConstant.java
@@ -1,0 +1,16 @@
+package atdd.user;
+
+public class TestConstant {
+  public static String USER_URL = "/user";
+  public static String USER_1_NAME = new String("손건");
+  public static String USER_1_EMAIL = new String("user1@gmail.com");
+  public static String USER_1_PASSWORD = new String("asdf");
+  
+  public static String USER_2_NAME = new String("손건2");
+  public static String USER_2_EMAIL = new String("user2@gmail.com");
+  public static String USER_2_PASSWORD = new String("asdf");
+
+  public static String USER_3_NAME = new String("손건3");
+  public static String USER_3_EMAIL = new String("user3@gmail.com");
+  public static String USER_3_PASSWORD = new String("asdf");
+}
