+++ src/test/java/atdd/user/web/UserManageAcceptanceTest.java
@@ -1,14 +1,14 @@
-package atdd.path.web;
+package atdd.user.web;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
 
-import static atdd.path.TestConstant.*;
+import static atdd.user.TestConstant.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import atdd.path.AbstractAcceptanceTest;
-import atdd.path.application.dto.UserResponseView;
+import atdd.AbstractAcceptanceTest;
+import atdd.user.application.dto.UserResponseView;
 
 public class UserManageAcceptanceTest extends AbstractAcceptanceTest {
   private UserManageHttpTest userManageHttpTest;
